+++ src/components/probe.rs
@@ -1,9 +1,9 @@
-use crate::common::{Component, Input, OutputType, Ports};
+use crate::common::{Component, Id, Input, OutputType, Ports};
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct Probe {
-    pub id: String,
+    pub id: Id,
     pub pos: (f32, f32),
     pub input: Input,
 }
@@ -14,16 +14,16 @@ impl Component for Probe {
         println!("Probe");
     }
 
-    fn get_id_ports(&self) -> (String, Ports) {
+    fn get_id_ports(&self) -> (Id, Ports) {
         (
             self.id.clone(),
-            Ports {
+            Ports::new(
                 // Probes take one input
-                inputs: vec![self.input.clone()],
-                out_type: OutputType::Combinatorial,
+                vec![&self.input],
+                OutputType::Combinatorial,
                 // No output value
-                outputs: vec![],
-            },
+                vec![],
+            ),
         )
     }
 }